+++ apps/vr-tests/src/stories/Spinner.stories.tsx
@@ -7,26 +7,16 @@ import { Spinner, SpinnerSize } from 'office-ui-fabric-react';
 
 storiesOf('Spinner', module)
   .addDecorator(FabricDecorator)
-  .addDecorator(story => (
-    <Screener
-      steps={new Screener.Steps()
-        .snapshot('default', { cropTo: '.testWrapper' })
-        .end()
-      }
-    >
-      {story()}
-    </Screener>
-  )).addStory('Extra small', () => (
-    <Spinner size={SpinnerSize.xSmall} />
-  )).addStory('Small', () => (
-    <Spinner size={SpinnerSize.small} />
-  )).addStory('Medium', () => (
-    <Spinner size={SpinnerSize.medium} />
-  )).addStory('Large', () => (
-    <Spinner size={SpinnerSize.large} />
-  )).addStory('Label', () => (
-    <Spinner
-      size={SpinnerSize.medium}
-      label='Spinner label'
-    />
+  .addDecorator(story => <Screener steps={new Screener.Steps().snapshot('default', { cropTo: '.testWrapper' }).end()}>{story()}</Screener>)
+  .addStory('Extra small', () => <Spinner size={SpinnerSize.xSmall} />)
+  .addStory('Small', () => <Spinner size={SpinnerSize.small} />)
+  .addStory('Medium', () => <Spinner size={SpinnerSize.medium} />)
+  .addStory('Large', () => <Spinner size={SpinnerSize.large} />)
+  .addStory('Label', () => <Spinner size={SpinnerSize.medium} label="Spinner label" />)
+  .addStory('Label at top', () => <Spinner size={SpinnerSize.medium} label="Spinner label" labelPosition="top" />)
+  .addStory('Label on the right', () => (
+    <Spinner size={SpinnerSize.medium} label="Spinner label" labelPosition="right" />
+  ))
+  .addStory('Label on the left', () => (
+    <Spinner size={SpinnerSize.medium} label="Spinner label" labelPosition="left" />
   ));