+++ packages/office-ui-fabric-react/etc/office-ui-fabric-react.api.ts
@@ -7464,6 +7464,16 @@ interface IDropdownSubComponentStyles {
   label: IStyleFunctionOrObject<ILabelStyleProps, any>;
 }
 
+// WARNING: Because this definition is explicitly marked as @internal, an underscore prefix ("_") should be added to its name
+// @internal
+interface IEffects {
+  elevation16: IRawStyle;
+  elevation4: IRawStyle;
+  elevation64: IRawStyle;
+  elevation8: IRawStyle;
+  roundedCorner2: number;
+}
+
 // @public
 interface IEntityRect {
   // (undocumented)
@@ -7957,7 +7967,7 @@ interface IGroupedList extends IList {
 }
 
 // @public (undocumented)
-interface IGroupedListProps extends React.Props<GroupedListBase> {
+interface IGroupedListProps extends React.ClassAttributes<GroupedListBase> {
   className?: string;
   compact?: boolean;
   componentRef?: IRefObject<IGroupedList>;
@@ -8225,7 +8235,7 @@ interface IKeytipLayer {
 }
 
 // @public (undocumented)
-interface IKeytipLayerProps extends React.Props<IKeytipLayer> {
+interface IKeytipLayerProps extends React.ClassAttributes<IKeytipLayer> {
   componentRef?: IRefObject<IKeytipLayer>;
   content: string;
   keytipExitSequences?: IKeytipTransitionKey[];
@@ -8623,7 +8633,7 @@ interface IModal {
 }
 
 // @public (undocumented)
-interface IModalProps extends React.Props<ModalBase>, IWithResponsiveModeState, IAccessiblePopupProps {
+interface IModalProps extends React.ClassAttributes<ModalBase>, IWithResponsiveModeState, IAccessiblePopupProps {
   className?: string;
   componentRef?: IRefObject<IModal>;
   containerClassName?: string;
@@ -8862,7 +8872,7 @@ interface IPage {
 }
 
 // @public (undocumented)
-interface IPageProps extends React.HTMLAttributes<HTMLDivElement>, React.Props<HTMLDivElement> {
+interface IPageProps extends React.HTMLAttributes<HTMLDivElement>, React.ClassAttributes<HTMLDivElement> {
   page: IPage;
   role?: string;
 }
@@ -9577,6 +9587,8 @@ interface IRGB {
 // @public (undocumented)
 interface IScheme {
   disableGlobalClassNames: boolean;
+  // (undocumented)
+  effects: IEffects;
   // (undocumented)
   fonts: IFontStyles;
   // (undocumented)
@@ -10323,13 +10335,14 @@ interface ISpinButtonStyles {
 interface ISpinner {
 }
 
-// @public (undocumented)
+// @public
 interface ISpinnerProps extends React.HTMLAttributes<HTMLElement> {
   ariaLabel?: string;
   ariaLive?: 'assertive' | 'polite' | 'off';
   className?: string;
   componentRef?: IRefObject<ISpinner>;
   label?: string;
+  labelPosition?: SpinnerLabelPosition;
   size?: SpinnerSize;
   styles?: IStyleFunctionOrObject<ISpinnerStyleProps, ISpinnerStyles>;
   theme?: ITheme;
@@ -10337,25 +10350,19 @@ interface ISpinnerProps extends React.HTMLAttributes<HTMLElement> {
   type?: SpinnerType;
 }
 
-// @public (undocumented)
+// @public
 interface ISpinnerStyleProps {
-  // (undocumented)
   className?: string;
-  // (undocumented)
+  labelPosition?: SpinnerLabelPosition;
   size?: SpinnerSize;
-  // (undocumented)
   theme: ITheme;
 }
 
-// @public (undocumented)
+// @public
 interface ISpinnerStyles {
-  // (undocumented)
   circle?: IStyle;
-  // (undocumented)
   label?: IStyle;
-  // (undocumented)
   root?: IStyle;
-  // (undocumented)
   screenReaderText?: IStyle;
 }
 
@@ -11959,7 +11966,7 @@ class SpinnerBase extends BaseComponent<ISpinnerProps, any> {
   render(): JSX.Element;
 }
 
-// @public (undocumented)
+// @public
 enum SpinnerSize {
   large = 3,
   medium = 2,
@@ -12616,6 +12623,7 @@ module ZIndexes {
 // WARNING: Unsupported export: Slider
 // WARNING: Unsupported export: ISliderStyleProps
 // WARNING: Unsupported export: Spinner
+// WARNING: Unsupported export: SpinnerLabelPosition
 // WARNING: Unsupported export: AnimationClassNames
 // WARNING: Unsupported export: FontClassNames
 // WARNING: Unsupported export: ColorClassNames